Move "assets" directory under "app/"
This matches the directory layout we have in other repos [1].

[1]: https://github.com/alphagov/document-download-frontend/tree/master/app/assets
