+++ app/assets/javascripts/govuk-frontend-details-init.mjs
