+++ app/assets/javascripts/govuk-frontend-details.mjs
