+++ app/assets/javascripts/relative-dates-init.mjs
