+++ app/assets/stylesheets/components/alert.scss
