+++ app/assets/stylesheets/components/alerts_icon.scss
