+++ app/assets/stylesheets/components/banner.scss
