+++ app/assets/stylesheets/components/example.scss
