+++ app/assets/stylesheets/components/image.scss
