+++ app/assets/stylesheets/components/responsive-embed.scss
